Change file descriptor to be blocking. (#4)
Force the interaction with the `trace_pipe` to be blocking, by setting the file descriptor setting `non-block` to `false`. This preserves behavior this library expects from the go runtime, and allows
it to be used in golang versions later than 1.8.

An alternative solution would be to redesign the polling loop to leverage the latest in golang IO capabilities, but that doesn't seem necessary at this time.